DevPage - Restrict route by permission
Router - add reference to store in the router, which the devpage route uses in it's beforeEnter function to check that the user has the route permission. If it gets specific value of true, then it routes there. If not, it redirects the routing to the NoAccessPage
DevPage - as a fall back, when the login is done loading, it checks for the route permission again ... and displays the page accordingly. This may not be required but I'm not sure how exploitable this permission access strategy is yet
-- add loading card to DevPage while login credentials are being checked
NoAccessPage - add page to display when user doesn't have correct permissions for a page or component but they've have attempted to reach them anyway
authModule, DA, api/User - add hasPermission and userHasPermission api call that checks for user permission with an 'action' (verb) and 'object' (noun)